[sourcegen] Split HeaderParser from clib
