+++ interfaces/sourcegen/sourcegen/clib/config.yaml
@@ -11,40 +11,10 @@ ignore_files: []
 ignore_funcs: {}
 
 # Return type cross-walks
-ret_type_crosswalk:
-  bool: int
-  void: int
-  int: int
-  size_t: int
-  double: double
-  const double: double
-  shared_ptr<T>: int
-  string: char*
-  const string: char*
-  vector<double>&: double*
-  const vector<double>&: double*
+ret_type_crosswalk: {}
 
 # Parameter type cross-walks
-prop_type_crosswalk:
-  bool: int
-  int: int
-  const int*: const int*
-  size_t: int
-  const size_t: int
-  double: double
-  const double: double
-  double*: double*
-  double* const: double*
-  const double*: const double*
-  const double* const: const double*
-  string: const char*
-  const string&: const char*
-  shared_ptr<T>: int
-  const shared_ptr<T>: int
-  vector<double>&: double*
-  const vector<double>&: const double*
-  vector<shared_ptr<T>>&: int[]
-  const vector<shared_ptr<T>>&: int[]
+prop_type_crosswalk: {}
 
 # Cabinets with associated preambles (headers)
 preambles:
