+++ interfaces/sourcegen/sourcegen/headers/__init__.py
@@ -1,0 +1,1 @@
+from .recipes import RecipeParser
