+++ interfaces/sourcegen/sourcegen/headers/config.yaml
@@ -1,0 +1,47 @@
+# Configuration for CLib header parser.
+
+# This file is part of Cantera. See License.txt in the top-level directory or
+# at https://cantera.org/license.txt for license and copyright information.
+
+# List of files to ignore entirely:
+ignore_files: []
+
+# Dictionary of file names and list of functions to ignore.
+# Example: ctkin_auto.yaml: [phase]
+ignore_funcs: {}
+
+# Return type cross-walks
+ret_type_crosswalk:
+  bool: int
+  void: int
+  int: int
+  size_t: int
+  double: double
+  const double: double
+  shared_ptr<T>: int
+  string: char*
+  const string: char*
+  vector<double>&: double*
+  const vector<double>&: double*
+
+# Parameter type cross-walks
+prop_type_crosswalk:
+  bool: int
+  int: int
+  const int*: const int*
+  size_t: int
+  const size_t: int
+  double: double
+  const double: double
+  double*: double*
+  double* const: double*
+  const double*: const double*
+  const double* const: const double*
+  string: const char*
+  const string&: const char*
+  shared_ptr<T>: int
+  const shared_ptr<T>: int
+  vector<double>&: double*
+  const vector<double>&: const double*
+  vector<shared_ptr<T>>&: int[]
+  const vector<shared_ptr<T>>&: int[]
